+++ src/main/java/Echo.java
@@ -47,120 +47,3 @@ public class Echo {
         System.out.println("____________________________________________________________");
     }
 }
-
-class Task {
-    private String description;
-    private boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public boolean isDone() {
-        return isDone;
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    public void markAsUndone() {
-        isDone = false;
-    }
-
-    @Override
-    public String toString() {
-        return "[" + (isDone ? "X" : " ") + "] " + description;
-    }
-}
-
-class TaskManager {
-    private List<Task> tasks;
-
-    public TaskManager() {
-        this.tasks = new ArrayList<>();
-    }
-
-    public void executeCommand(String command) {
-        String[] tokens = command.split(" ", 2);
-
-        switch (tokens[0].toLowerCase()) {
-            case "list":
-                listTasks();
-                break;
-            case "mark":
-                markTask(tokens);
-                break;
-            case "unmark":
-                unmarkTask(tokens);
-                break;
-            default:
-                addTask(command);
-        }
-    }
-
-    private void listTasks() {
-        System.out.println("____________________________________________________________");
-        if (tasks.isEmpty()) {
-            System.out.println("No tasks in the list.");
-        } else {
-            System.out.println("Here are the tasks in your list:");
-            for (int i = 0; i < tasks.size(); i++) {
-                System.out.println((i + 1) + ". " + tasks.get(i));
-            }
-        }
-        System.out.println("____________________________________________________________");
-    }
-
-    private void markTask(String[] tokens) {
-        if (tokens.length == 2) {
-            int index = Integer.parseInt(tokens[1]);
-            if (isValidIndex(index)) {
-                tasks.get(index - 1).markAsDone();
-                System.out.println("____________________________________________________________");
-                System.out.println("Nice! I've marked this task as done:");
-                System.out.println("  " + tasks.get(index - 1));
-                System.out.println("____________________________________________________________");
-            } else {
-                System.out.println("Invalid task index.");
-            }
-        } else {
-            System.out.println("Invalid command. Usage: mark <index>");
-        }
-    }
-
-    private void unmarkTask(String[] tokens) {
-        if (tokens.length == 2) {
-            int index = Integer.parseInt(tokens[1]);
-            if (isValidIndex(index)) {
-                tasks.get(index - 1).markAsUndone();
-                System.out.println("____________________________________________________________");
-                System.out.println("OK, I've marked this task as not done yet:");
-                System.out.println("  " + tasks.get(index - 1));
-                System.out.println("____________________________________________________________");
-            } else {
-                System.out.println("Invalid task index.");
-            }
-        } else {
-            System.out.println("Invalid command. Usage: unmark <index>");
-        }
-    }
-
-    private void addTask(String description) {
-        if (!description.equals("bye")) {
-            tasks.add(new Task(description));
-            System.out.println("____________________________________________________________");
-            System.out.println("added: " + description);
-            System.out.println("____________________________________________________________");
-        }
-    }
-
-    private boolean isValidIndex(int index) {
-        return index >= 1 && index <= tasks.size();
-    }
-}
